requirements.txt
numpy==1.15.1
Pillow==5.2.0
